WIP - Move "Domains" Exceptions to "Api"
