+++ src/Exception/Api/UnauthorizedException.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+namespace CarAndClassic\TalkJS\Exception\Api;
+
+use CarAndClassic\TalkJS\Exception\ApiException;
+
+final class UnauthorizedException extends \Exception implements ApiException
+{
+}
