--- src/Exception/Domain/UnauthorizedException.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This software may be modified and distributed under the terms
- * of the MIT license. See the LICENSE file for details.
- */
-
-namespace CarAndClassic\TalkJS\Exception\Domain;
-
-use CarAndClassic\TalkJS\Exception\DomainException;
-
-final class UnauthorizedException extends \Exception implements DomainException
-{
-}
